1_model/requirements.txt
langchain
langchain-huggingface
langchain-openai
langchain-community
langchain-anthropic
langchain-google-genai
openai
google-generativeai
transformers
huggingface-hub
python-dotenv
numpy
scikit-learn
pandas
torch 
torchvision 
torchaudio
